requirements.txt
albumentations==1.3.0
argparse
configparser
distinctipy
easydict
fairscale
faiss-gpu
h5py
imageio
grad-cam
matplotlib
numpy
numpy-quaternion
Ninja
OmegaConf
open3d
opencv-python
pandas
prettytable
pyyaml
scikit-learn
scipy
shapely
shutilwhich
tensorboard
termcolor
timm==0.9.2
#torch
torchvision
tqdm